+++ monticore-grammar/src/main/java/de/monticore/expressions/lambdaexpressions/interpreter/LambdaExpressionsInterpreter.java
@@ -9,7 +9,7 @@ import de.monticore.interpreter.calculations.MICalculation;
 import de.monticore.interpreter.calculations.MICalculationValue;
 import de.monticore.symbols.basicsymbols.interpreter.frames.MIFrameLayoutForBasicSymbols;
 import de.monticore.interpreter.setters.MISetter;
-import de.monticore.interpreter.util.InterpreterData;
+import de.monticore.interpreter.util.InterpreterDataForBasicSymbols;
 import de.monticore.interpreter.values.MCValueFunctionOfModel;
 import de.monticore.symbols.basicsymbols._symboltable.VariableSymbol;
 
@@ -21,9 +21,9 @@ import java.util.List;
 public class LambdaExpressionsInterpreter
     extends LambdaExpressionsInheritanceHandler {
 
-  protected InterpreterData iData;
+  protected InterpreterDataForBasicSymbols iData;
 
-  public LambdaExpressionsInterpreter(InterpreterData iData) {
+  public LambdaExpressionsInterpreter(InterpreterDataForBasicSymbols iData) {
     this.iData = Preconditions.checkNotNull(iData);
   }
