+++ monticore-grammar/src/main/java/de/monticore/interpreter/signals/MCSignalContinue.java
@@ -1,0 +1,13 @@
+package de.monticore.interpreter.signals;
+
+public class MCSignalContinue
+    extends MCSignalFlowControl {
+
+  /**
+   * signals the current thread.
+   */
+  public static void signal() {
+    throw new MCSignalContinue();
+  }
+
+}
