+++ monticore-grammar/src/main/java/de/monticore/statements/mcreturnstatements/interpreter/MCReturnStatementsInterpreter.java
@@ -4,8 +4,8 @@ package de.monticore.statements.mcreturnstatements.interpreter;
 import com.google.common.base.Preconditions;
 import de.monticore.interpreter.calculations.MICalculationValue;
 import de.monticore.interpreter.calculations.MICalculationVoid;
-import de.monticore.interpreter.util.InterpreterData;
-import de.monticore.interpreter.values.MCSignalReturnForInterpreter;
+import de.monticore.interpreter.util.InterpreterDataForBasicSymbols;
+import de.monticore.interpreter.signals.MCSignalReturn;
 import de.monticore.statements.mcreturnstatements._ast.ASTReturnStatement;
 import de.monticore.statements.mcreturnstatements._visitor.MCReturnStatementsInheritanceHandler;
 
@@ -15,9 +15,9 @@ import de.monticore.statements.mcreturnstatements._visitor.MCReturnStatementsInh
 public class MCReturnStatementsInterpreter
     extends MCReturnStatementsInheritanceHandler {
 
-  protected InterpreterData iData;
+  protected InterpreterDataForBasicSymbols iData;
 
-  public MCReturnStatementsInterpreter(InterpreterData iData) {
+  public MCReturnStatementsInterpreter(InterpreterDataForBasicSymbols iData) {
     this.iData = Preconditions.checkNotNull(iData);
   }
 
@@ -29,11 +29,11 @@ public class MCReturnStatementsInterpreter
       MICalculationValue exprCalc =
           iData.popCalculation().asCalculationValue();
       returnSignalCalc = frame ->
-          MCSignalReturnForInterpreter.signal(exprCalc.calculate(frame));
+          MCSignalReturn.signal(exprCalc.calculate(frame));
     }
     else {
       returnSignalCalc = frame ->
-          MCSignalReturnForInterpreter.signal();
+          MCSignalReturn.signal();
     }
     iData.putCalculation(returnSignalCalc);
   }
