+++ monticore-grammar/src/test/java/de/monticore/interpreter/util/InterpreterAccess4Tests.java
@@ -2,26 +2,26 @@
 package de.monticore.interpreter.util;
 
 import de.monticore.ast.ASTNode;
+import de.monticore.expressions.interpreter.ExpressionsInterpreter;
 import de.monticore.interpreter.calculations.MICalculation;
 import de.monticore.values.MCValue;
-import de.monticore.statements.interpreter.StatementsInterpreter;
 import de.monticore.visitor.ITraverser;
 
 /**
  * To access the internals
  */
-public class InterpreterAccess4Tests extends StatementsInterpreter {
+public class InterpreterAccess4Tests extends ExpressionsInterpreter {
 
-  public InterpreterAccess4Tests(TraverserAndIData interpreterTraverser) {
-    super(interpreterTraverser);
+  public InterpreterAccess4Tests(ITraverser traverser, InterpreterDataForBasicSymbols iData) {
+    super(traverser, iData);
   }
 
   public ITraverser getTraverser() {
-    return interpreterTraverser.traverser();
+    return traverser;
   }
 
-  public InterpreterData getInterpreterData() {
-    return interpreterTraverser.data();
+  public InterpreterDataForBasicSymbols getInterpreterData() {
+    return iData;
   }
 
   @Override
