--- monticore-runtime/src/main/java/de/monticore/values/MCSignalContinue.java
@@ -1,20 +1,0 @@
-package de.monticore.values;
-
-public interface MCSignalContinue extends MCSignalFlowControl {
-
-  @Override
-  default boolean isContinue() {
-    return true;
-  }
-
-  @Override
-  default String printType() {
-    return "Continue";
-  }
-
-  @Override
-  default String printValue() {
-    return "";
-  }
-
-}
